pyproject.toml
[tool.poetry]
name = "rlhabitt_backend"
version = "0.1.0"
description = "Download shared files from OneDrive using device-based authentication"
authors = ["User"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
msal = "^1.32.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
openpyxl = "^3.1.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
onedrive-download = "src.__main__:main"
